# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Rental App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1323",
    "paths": {
...
        "/calendar": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "To do create new calendar event",
                "parameters": [
                    {
                        "description": "json req body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateCalendar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/calendar/timeline": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "To do get data timeline",
                "parameters": [
                    {
                        "description": "json req body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Timeline"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/response.Timeline"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/calendar/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "To do get a calendar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Calendar ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/response.Calendar"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "To do update a calendar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Calendar ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json req body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateCalendar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Calendar"
                ],
                "summary": "Delete Calendar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Calendar ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.CalendarView": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "string"
                },
                "createBy": {
                    "type": "string"
                },
                "createDt": {
                    "type": "string"
                },
                "createName": {
                    "type": "string"
                },
                "deleteBy": {
                    "type": "string"
                },
                "deleteDt": {
                    "type": "string"
                },
                "deleteName": {
                    "type": "string"
                },
                "endDt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "propertyId": {
                    "type": "string"
                },
                "propertyName": {
                    "type": "string"
                },
                "startDt": {
                    "type": "string"
                },
                "updateBy": {
                    "type": "string"
                },
                "updateDt": {
                    "type": "string"
                },
                "updateName": {
                    "type": "string"
                }
            }
        },
        "model.PropertyView": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "string"
                },
                "createBy": {
                    "type": "string"
                },
                "createDt": {
                    "type": "string"
                },
                "createName": {
                    "type": "string"
                },
                "deleteBy": {
                    "type": "string"
                },
                "deleteDt": {
                    "type": "string"
                },
                "deleteName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updateBy": {
                    "type": "string"
                },
                "updateDt": {
                    "type": "string"
                },
                "updateName": {
                    "type": "string"
                }
            }
        },
        "request.CreateCalendar": {
            "type": "object",
            "required": [
                "endDt",
                "name",
                "propertyId",
                "startDt"
            ],
            "properties": {
                "endDt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "propertyId": {
                    "type": "string"
                },
                "startDt": {
                    "type": "string"
                }
            }
        },
...
        "request.Timeline": {
            "type": "object",
            "required": [
                "companyId",
                "endDt",
                "startDt"
            ],
            "properties": {
                "companyId": {
                    "type": "string"
                },
                "endDt": {
                    "type": "string"
                },
                "startDt": {
                    "type": "string"
                }
            }
        },
        "request.UpdateCalendar": {
            "type": "object",
            "required": [
                "endDt",
                "name",
                "propertyId",
                "startDt"
            ],
            "properties": {
                "endDt": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "propertyId": {
                    "type": "string"
                },
                "startDt": {
                    "type": "string"
                }
            }
        },
...
        "response.Calendar": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "string"
                },
                "createBy": {
                    "type": "string"
                },
                "createDt": {
                    "type": "string"
                },
                "createName": {
                    "type": "string"
                },
                "deleteBy": {
                    "type": "string"
                },
                "deleteDt": {
                    "type": "string"
                },
                "deleteName": {
                    "type": "string"
                },
                "endDt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "propertyId": {
                    "type": "string"
                },
                "propertyName": {
                    "type": "string"
                },
                "startDt": {
                    "type": "string"
                },
                "updateBy": {
                    "type": "string"
                },
                "updateDt": {
                    "type": "string"
                },
                "updateName": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "response.Timeline": {
            "type": "object",
            "properties": {
                "calendars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CalendarView"
                    }
                },
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PropertyView"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}